Renamespace ready button
