+++ osu.Game.Tests/Visual/RealtimeMultiplayer/TestSceneRealtimeReadyButton.cs
@@ -11,7 +11,7 @@ using osu.Game.Beatmaps;
 using osu.Game.Online.Multiplayer;
 using osu.Game.Online.RealtimeMultiplayer;
 using osu.Game.Rulesets;
-using osu.Game.Screens.Multi.RealtimeMultiplayer;
+using osu.Game.Screens.Multi.RealtimeMultiplayer.Match;
 using osu.Game.Tests.Resources;
 using osu.Game.Users;
 using osuTK;
